(add) express router implmeneted routes split into techs and logs files in routes folder, server.js moved to src/server file
