+++ package.json
@@ -26,7 +26,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "server": "nodemon server.js"
+    "server": "nodemon ./src/server/server.js"
   },
   "eslintConfig": {
     "extends": [
